+++ src/components/admin/EditEventForm/style.less
@@ -1,5 +1,5 @@
-@import '../../styles/fonts.less';
-@import '../../styles/vars.less';
+@import '../../../styles/fonts.less';
+@import '../../../styles/vars.less';
 
 .edit-event-form {
   display: flex;
