+++ src/components/admin/EditEventPage/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import EditEventForm from '../../containers/admin/EditEventForm';
+import EditEventForm from '../../../containers/admin/EditEventForm';
 
 import './style.less';
