+++ src/components/admin/EditEventPage/style.less
@@ -1,4 +1,4 @@
-@import '../../styles/vars.less';
+@import '../../../styles/vars.less';
 
 .edit-event-page {
   .title {
